+++ test/fools.js
@@ -4,134 +4,68 @@ var _ = require('underscore');
 
 describe('Fools', function () {
 
-    describe('loop', function () {
+    /**
+     * Rate evaluates candidates based on preprogrammed properties and weights.
+     *
+     */
 
-        describe ('algorithmic', function(){
+    describe('rate', function () {
 
-            var scan, looper;
+        var students, math_score, gpa, gpa_flat;
 
-            before(function(){
+        before(function () {
+            students = [
+                {name: 'bob', math: 2.0, english: 3.0, pe: 4.0, health: 3.0},
+                {name: 'stan', math: 4.0, english: 4.0, pe: 2.0, health: 4.0},
+                {name: 'rick', math: 3.5, english: 3.6, pe: 4.0, health: 4.0},
+                {name: 'sue', math: 4.0, english: 4.0, pe: 3.5 , health: 3.5}
+            ];
+            math_score = Fools.rate().prop('math');
 
-                looper = Fools.loop(function(iterator, memo, out){
-                 //   console.log('iterator: %s', require('util').inspect(iterator));
-                    switch(out.place('x', iterator)){
-                        case 'first':
-                            memo.push( iterator.x);
-                            break;
+            gpa_flat = Fools.rate();
+            gpa_flat.property('math');
+            gpa_flat.property('english');
+            gpa_flat.property('pe');
+            gpa_flat.property('health');
 
-                        case 'last':
-                            memo.push( iterator.x);
-                            break;
-
-                        default:
-                            switch(out.place('y', iterator)){
-                                case 'first':
-                                    memo.push(iterator.y);
-                                    break;
-
-                                case 'last':
-                                    memo.push(iterator.y);
-                                    break;
-
-                                default:
-                                    memo.push(0);
-                            }
-                    }
-                    return memo;
-                }).dim('x')
-                    .min(function(){ return scan.x - scan.width})
-                    .max(function(){return scan.x + scan.width})
-                .dim('y')
-                    .min(function(){ return scan.y - scan.height})
-                    .max(function(){return scan.y + scan.height});
-
-
-            });
-
-
-            it('should reflect scans position', function(){
-                scan = {width: 1, height: 2, x: 0, y: 0};
-                looper([]).should.eql([ -1, -2, 1, -1, 0, 1, -1, 0, 1, -1, 0, 1, -1, 2, 1 ]);
-            });
-
-            it('should reflect scans updated position', function(){
-                scan = {width: 3, height: 4, x:2, y: 2};
-                looper([]).should.eql([ -1, -2, -2, -2, -2, -2, 5, -1, 0, 0, 0, 0, 0, 5, -1, 0, 0, 0, 0, 0, 5, -1, 0, 0, 0, 0, 0, 5, -1, 0, 0, 0, 0, 0, 5, -1, 0, 0, 0, 0, 0, 5, -1, 0, 0, 0, 0, 0, 5, -1, 0, 0, 0, 0, 0, 5, -1, 6, 6, 6, 6, 6, 5 ]);
-            })
+            gpa = Fools.rate();
+            gpa.property('math').weight(4);
+            gpa.property('english').weight(3);
+            gpa.property('pe').weight(2);
+            gpa.property('health').weight(3);
         });
 
-        describe('place', function () {
+        it('should rate students on math', function () {
+            math_score.rate(students[0]).should.eql(2, 'bob has a 2 math');
+            math_score._candidates.length.should.eql(0, 'no candidates added');
 
-            var looper;
+            math_score.add(students);
+            math_score._candidates.length.should.eql(4, '4 added records');
+            math_score.best().data.name.should.eql('sue');
+            math_score.worst().data.name.should.eql('bob');
 
-            before(function () {
+            gpa_flat.rate(students[0]).should.eql(3.0, 'bobs flat gpa == 3.0');
 
-                looper = Fools.loop(function (iterator, memo, out) {
-                    memo.push(out.place('x', iterator)
-                        + '...' + out.place('y', iterator));
-                    return memo;
-                }).dim('x').min(0).max(3).dim('y').min(0).max(3);
+           // var g = gpa.rate(students[2]);
+          //  console.log('gpa grade', g);
+            gpa.rate(students[0]).should.be.approximately(2.83, 0.01, 'weighted grade for bob -- hurts on the fundamentals');
+            gpa.rate(students[1]).should.be.approximately(3.66, 0.01, 'weighted grade for stan -- low PE score doesnt hurt too much');
+            gpa.rate(students[2]).should.be.approximately(3.73, 0.01, 'weighted grade for rick -- good but in the wrong things');
+            gpa.rate(students[3]).should.be.approximately(3.79, 0.01, 'weighted grade for sue -- high overall');
 
-            });
+          //  var g = gpa_flat.rate(students[2]);
+        //   console.log('gpa_flat grade', g);
+            gpa_flat.rate(students[0]).should.be.approximately(3, 0.01, 'weighted grade for bob -- hurts on the fundamentals');
+            gpa_flat.rate(students[1]).should.be.approximately(3.5, 0.01, 'weighted grade for stan -- low PE score doesnt hurt too much');
+            gpa_flat.rate(students[2]).should.be.approximately(3.77, 0.01, 'weighted grade for rick -- overall best');
+            gpa_flat.rate(students[3]).should.be.approximately(3.75, 0.01, 'weighted grade for sue -- good but not overall best');
 
-            it('should be funky', function () {
-                looper([]).should.eql(['first...first',
-                    'middle...first',
-                    'middle...first',
-                    'last...first',
-                    'first...middle',
-                    'middle...middle',
-                    'middle...middle',
-                    'last...middle',
-                    'first...middle',
-                    'middle...middle',
-                    'middle...middle',
-                    'last...middle',
-                    'first...last',
-                    'middle...last',
-                    'middle...last',
-                    'last...last']);
-            });
+            gpa_flat.add(students);
+            gpa.add(students);
 
-        });
-
-        describe('basic', function () {
-
-            var looper;
-
-            before(function () {
-                looper = Fools.loop(function (iterator, memo) {
-
-                    if ((!(iterator.x % 3)) && (!(iterator.y % 3))) {
-                        memo.push(2);
-                    } else if ((!(iterator.x % 3)) || (!(iterator.y % 3))) {
-                        memo.push(1);
-                    } else {
-                        memo.push(0);
-                    }
-
-                    return memo;
-
-                }).dim('x').min(0).max(5).dim('y').min(0).max(6);
-            });
-
-            it('should have expected dims', function () {
-                looper.dims.should.eql({ x: { min: 0, max: 5 }, y: { min: 0, max: 6 } });
-            });
-
-            it('should produce data', function () {
-                var data = looper([]);
-                data.should.eql(
-                    [2, 1, 1, 2, 1, 1,
-                        1, 0, 0, 1, 0, 0,
-                        1, 0, 0, 1, 0, 0,
-                        2, 1, 1, 2, 1, 1,
-                        1, 0, 0, 1, 0, 0,
-                        1, 0, 0, 1, 0, 0,
-                        2, 1, 1, 2, 1, 1], 'array produced');
-            });
+            gpa_flat.best().data.name.should.eql('rick', 'rick is the best flat gpa');
+            gpa.best().data.name.should.eql('sue', 'sue is the best weighted gpa');
         })
-
     });
 
     describe('until', function () {
@@ -324,7 +258,7 @@ describe('Fools', function () {
                     });
 
                 if_positive.run(1).run(0).run(5).run(-1).run('foo');
-            })
+            });
 
             it('should tally results', function () {
                 positive.should.equal(3, 'three positive numbers');
@@ -406,9 +340,7 @@ describe('Fools', function () {
                 }).then(function (text) {
                         var m = grep.exec(text);
                         return parseFloat(m[1]);
-                    }).else(function () {
-                        return 0;
-                    }).err(function (err) {
+                    }).else(0).err(function (err) {
                         ++bad;
                         return 0;
                     });
@@ -432,9 +364,9 @@ describe('Fools', function () {
             before(function () {
                 var make_member = function (id, name, likes) {
                     var m = {
-                        id: id,
-                        name: name,
-                        likes: likes,
+                        id:         id,
+                        name:       name,
+                        likes:      likes,
                         likability: 0
                     };
 
@@ -533,7 +465,7 @@ describe('Fools', function () {
         var negative_numbers = [];
 
         var range_fn = Fools.range();
-        range_fn.filter(Math.abs);
+        range_fn.add_filter(Math.abs);
 
         var tally_negative = Fools.fork(function (n) {
             return n >= 0
@@ -576,15 +508,19 @@ describe('Fools', function () {
         it('should have these ones', function () {
             _s(ones).should.eql([-1, 0, 1, 4]);
         });
+
         it('should have these tens', function () {
             _s(tens).should.eql(_s([ 10, 20, 30]));
         });
+
         it('should have these hundreds', function () {
             _s(hundreds).should.eql(_s([500, -200, -100]));
         });
+
         it('should have these thousands', function () {
             _s(thousands).should.eql(_s([1000, -5000]));
         });
+
         it('should have these huge', function () {
             _s(huge).should.eql(_s([-100000, 2000000, 100235252]));
         });
@@ -592,6 +528,7 @@ describe('Fools', function () {
         it('should have these positive numbers', function () {
             _s(positive_numbers).should.eql(_s([1, 10, 20, 30, 0, 4, 1000, 500, 2000000, 100235252]));
         });
+
         it('should have these negative_numbers', function () {
             _s(negative_numbers).should.eql(_s([-1, -200, -100, -5000, -100000]));
         });