+++ CHANGELOG.md
@@ -1,2 +1,13 @@
-## 0.0.5
-Adding looping abstract
+# Changelog
+
+# Base
+
+Established fork, pipe, range, and until; tested
+
+# 0.0.5
+
+Added rate
+
+# 0.0.7
+
+Allowed filter to throw error in range